Outsourced menu items into separate components
